Added support for providing className to elements using Element.attr('class', 'your-class-name');
The changes were in both VML and SVG renderer where within setFillAndStroke check for class parameter and apply to element. For VML, it is applied as className property of node. The only extra computation comes in case where class is applied to group. The default class of the group is also retained.

The only change in core.js is to add class as part of availableAttrs