RedRaphael whenever required by a library in a non AMD fashion and AMD based module loader is used, RedRaphael is not itegrated properly.
Whenever a module loader is used in the page and RedRaphael detects it, RedRaphael defines itself in modules.
But if the library integrating RedRaphael doesn't require the RedRaphael modules in the module definition fashion, RedRaphael will never be integrated.

Fix is to use a flag based approach to prevent RedRaphael from defining itslef into modules.

The flag used is `optOutModuleLoading`.

Changes are also made in GruntFile.js `build` task to accomodate the changes where the SVG and VML modules are added before `//EXPOSE` line, the regex to determine
that is rectified.