Fix flow of multiline text including RTL characters. Currently fixed for Webkit based browsers.
Tested primarily on current versions of Chrome, Safari, Opera and Firefox (no issue). The issue in IE is yet to be resolved. The fix doesn't require dir="rtl" to be set on container or use of <bdo> to enable proper flow of the text with RTL characters. The unicode-bidi algorithm of the browsers handle them as required. The fix here breaks the continuity of successive <tspan> elements for text with RTL characters to be displayed in multiline. This is done by injecting dummy <tspan> element with LTR character, in between two consecutive <tspan>  s with RTL text.